+++ src/main/java/es/in2/wallet/application/dto/CredentialResponse.java
@@ -2,28 +2,20 @@ package es.in2.wallet.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
 import java.util.List;
 
-@Schema(description = "Implements the credential response according to " +
-        "https://github.com/hesusruiz/EUDIMVP/blob/main/issuance.md#credential-response")
 @Builder
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record CredentialResponse(
-
-        @JsonProperty("credentials") List<Credentials> credentials,
+        @JsonProperty(value = "credentials") List<Credential> credentials,
         @JsonProperty("transaction_id") String transactionId,
-        @JsonProperty("notification_id") String notificationId
-) {
-        @Builder
-        public record Credentials(
-                @JsonProperty("credential") String credential
+        @JsonProperty("interval") Long interval,
+        @JsonProperty("notification_id") String notificationId) {
 
-        ) {
-        }
+    @Builder
+    public record Credential(@JsonProperty String credential) {
+
+    }
 }
-
-
-