+++ src/main/java/es/in2/wallet/application/dto/TokenInfo.java
@@ -1,0 +1,7 @@
+package es.in2.wallet.application.dto;
+
+import lombok.Builder;
+
+@Builder
+public record TokenInfo(String accessToken, String refreshToken, long tokenObtainedAt, long expiresIn) {
+}